+++ applications/lokqlDx/Messages/DisplayResultMessage.cs
@@ -1,0 +1,4 @@
+﻿using CommunityToolkit.Mvvm.Messaging.Messages;
+using LokqlDx.ViewModels;
+
+public class DisplayResultMessage(NamedKustoResult result) : ValueChangedMessage<NamedKustoResult>(result);
