+++ applications/lokqlDx/Messages/QueryResultWithSender.cs
@@ -1,0 +1,3 @@
+﻿using KustoLoco.Core;
+
+public readonly record struct QueryResultWithSender(string Sender, KustoQueryResult Result, bool ImmediateDisplay);
