+++ applications/lokqlDx/Messages/SchemaUpdateMessage.cs
@@ -1,0 +1,4 @@
+﻿using CommunityToolkit.Mvvm.Messaging.Messages;
+using Lokql.Engine;
+
+public class SchemaUpdateMessage(SchemaLine[] lines) : ValueChangedMessage<SchemaLine[]>(lines);
