+++ applications/lokqlDx/Messages/ShowQueryRequestMessage.cs
@@ -1,0 +1,4 @@
+﻿using CommunityToolkit.Mvvm.Messaging.Messages;
+using LokqlDx.ViewModels;
+
+public class ShowQueryRequestMessage(QueryDocumentViewModel query) : ValueChangedMessage<QueryDocumentViewModel>(query);
