+++ applications/lokqlDx/Messages/ShowToolMessage.cs
@@ -1,0 +1,3 @@
+﻿using CommunityToolkit.Mvvm.Messaging.Messages;
+
+public class ShowToolMessage(string tool) : ValueChangedMessage<string>(tool);
