+++ applications/lokqlDx/Messages/ThemeChangedMessage.cs
@@ -1,0 +1,3 @@
+﻿using CommunityToolkit.Mvvm.Messaging.Messages;
+
+public class ThemeChangedMessage(string theme) : ValueChangedMessage<string>(theme);
